+++ devices/lock.js
@@ -5,7 +5,7 @@ class lockMechanism {
 	constructor(platform, log) {
 		this.log = log
 		this.platform = platform
-		this.wallboxapi = new wallboxAPI(this, log)
+		this.wallboxapi = new wallboxAPI(this.platform, log)
 	}
 
 	createLockAccessory(device, config, uuid, platformAccessory) {