+++ devices/outlet.js
@@ -5,7 +5,7 @@ class basicOutlet {
 	constructor(platform, log, config) {
 		this.log = log
 		this.platform = platform
-		this.wallboxapi = new wallboxAPI(this, log)
+		this.wallboxapi = new wallboxAPI(this.platform, log)
 		this.enumeration = enumeration
 	}
 
@@ -31,6 +31,8 @@ class basicOutlet {
 	}
 
 	async setOutletValue(device, outletService, value, callback) {
+		let statusCode
+		let currentMode
 		if (outletService.getCharacteristic(Characteristic.StatusFault).value == Characteristic.StatusFault.GENERAL_FAULT) {
 			callback('error')
 		}
@@ -40,7 +42,7 @@ class basicOutlet {
 			try {
 				statusCode = chargerData.status
 				currentMode = this.enumeration.items.filter(result => result.status == statusCode)[0].mode
-				this.log.debug('checking statuscod = %s, current mode = %s', statusCode, currentMode)
+				this.log.debug('checking status code = %s, current mode = %s', statusCode, currentMode)
 			} catch (error) {
 				statusCode = 'unknown'
 				currentMode = 'unknown'